pyproject.toml
[project]
name = "telegram_bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiogram>=3.23.0",
    "aiosqlite>=0.22.1",
    "python-dotenv>=1.2.1",
    "sqlalchemy>=2.0.45",
]
